Add FilterUnconnected Renderer, wire up Process and ProcessName Renderers.
